requirements.txt
apache-flink
pyyaml
requests
ollama          # ← Only AI calls
pydantic
fastapi
uvicorn
python-multipart
